+++ normandy/recipes/tests/__init__.py
@@ -1,15 +1,7 @@
 import factory
-from factory import fuzzy
 
+from normandy.base.tests import FuzzyUnicode
 from normandy.recipes.models import Action, Recipe, RecipeAction
-
-
-class FuzzyUnicode(fuzzy.FuzzyText):
-    """A FuzzyText factory that contains at least one non-ASCII character."""
-
-    def __init__(self, prefix=u'', **kwargs):
-        prefix = '%sđ' % prefix
-        super(FuzzyUnicode, self).__init__(prefix=prefix, **kwargs)
 
 
 class RecipeFactory(factory.DjangoModelFactory):
